ai_services/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-dotenv # For .env file loading for config
